Make the export a named export
This changes the export of the control to be a named export `SurveyAd`
rather than be the `module.exports` value. This is because bit-docs
greedily calls all functions that are exported by plugins. Making this
be a named export works around that issue.

Note that this is a breaking change, so downstream users will need to be
updated (next time they need to upgrade their survey-ad version that
		is).